pyproject.toml
[project]
name = "wa-agent-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "agentic>=0.0.2",
    "agentic-chunker>=0.0.6",
    "aiohappyeyeballs==2.6.1",
    "aiohttp>=3.11.0",
    "aiosignal==1.4.0",
    "alembic==1.16.5",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "authlib==1.6.3",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.5",
    "build==1.3.0",
    "cachetools==5.5.2",
    "camelot-py>=0.11.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "chromadb>=1.5.5",
    "chunker>=1.1",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "crewai[tools]>=0.95.0",
    "cryptography==45.0.7",
    "cyclopts==3.24.0",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "deepagents>=0.0.5",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22",
    "durationpy==0.10",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastapi==0.116.1",
    "fastmcp==2.12.3",
    "filelock==3.19.1",
    "filetype==1.2.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "ftfy==6.3.1",
    "ghostscript>=0.8.1",
    "google==3.0.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core>=2.25.1",
    "google-api-python-client==2.181.0",
    "google-apps-meet==0.1.16",
    "google-auth>=2.45.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "groq==0.31.1",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.1.9",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==6.30.1",
    "ipython==8.37.0",
    "isodate==0.7.2",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.2",
    "jpype1>=1.6.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kubernetes==33.1.0",
    "lance-namespace==0.0.6",
    "lance-namespace-urllib3-client==0.0.14",
    "lancedb==0.25.0",
    "langchain>=1.2.13",
    "langchain-cerebras>=0.8.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langfuse>=3.6.2",
    "langgraph>=1.1.1",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-prebuilt>=1.0.8",
    "langgraph-sdk>=0.3.12",
    "langsmith==0.4.21",
    "lazy-object-proxy==1.12.0",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mcp==1.14.0",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "ngrok==1.5.1",
    "numpy==2.3.2",
    "nvidia-ml-py3>=7.352.0",
    "oauthlib==3.3.1",
    "ollama==0.5.3",
    "onnxruntime==1.22.1",
    "open-clip-torch==3.1.0",
    "openai>=2.30.0",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "opencv-python>=4.11.0.86",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-otlp-proto-common==1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.36.0",
    "opentelemetry-proto==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "orjson>=3.11.5",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "parse==1.20.2",
    "parso==0.8.5",
    "pathable==0.4.4",
    "pathlib-mate==1.3.2",
    "pdfplumber>=0.11.7",
    "pillow==11.3.0",
    "pip==25.2",
    "platformdirs==4.3.8",
    "portalocker==2.10.1",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "psycopg2==2.9.10",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pycparser==2.22",
    "pydantic>=2.10.0",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pylance==0.35.0",
    "pyodbc>=5.3.0",
    "pyparsing==3.2.4",
    "pypdf==6.0.0",
    "pypdf2>=3.0.1",
    "pyperclip==1.9.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyprojroot>=0.3.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-http-client==3.3.7",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.2",
    "qdrant-client==1.15.1",
    "referencing==0.36.2",
    "regex==2025.8.29",
    "requests>=2.30.0",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rich==13.9.4",
    "rich-rst==1.3.1",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "sendgrid==6.12.4",
    "sentence-transformers==5.1.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "stack-data==0.6.3",
    "starlette==0.47.0",
    "structlog==23.3.0",
    "sympy==1.14.0",
    "tabula-py>=2.10.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "timm==1.0.19",
    "tokenizers==0.22.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.56.0",
    "typer==0.16.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "unstructured[all-docs]>=0.18.15",
    "uritemplate==4.2.0",
    "urllib3>=2.3.0",
    "uvicorn==0.35.0",
    "vaul==1.1.1",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "werkzeug==3.1.1",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.24.0",
]
